--- src/components/globalErorModal.tsx
@@ -1,52 +1,0 @@
-import { useEffect } from "react";
-import GlobalModalDialog from "./globalModalDialog";
-import { Button, Dialog } from "@mui/material";
-import { PALLETE } from "../config/config";
-import { Height } from "@mui/icons-material";
-
-const GlobalErorrModal = (props: any) => {
-   interface GlobalErorrModalProps {
-      // showError: boolean;
-      onClose: () => void;
-    }
-    const handleClickClose = () => {
-      props.onClose();
-    };
-  useEffect(() => {}, []);
-  return (
-    <Dialog open={true} fullWidth sx={{maxHeight:"60vh",top:"13vh"}}>
-      <GlobalModalDialog
-        title={"Erorr"}
-        isButton={true}
-        img={"error.svg"}
-        children={
-            <div>
-              <Button
-                style={{
-                  zIndex: 3,
-                  position: "absolute",
-                  backgroundColor: PALLETE.ORANGE,
-                  borderRadius: "10px",
-                  top: "45vh",
-                  right: "4vw",
-                  color: "black",
-                }}
-                onClick={handleClickClose}
-              >
-                close
-              </Button><h2>
-              unexpected error
-              <br/> please try again later
-              <br/>
-              or contact our support team at:
-              <br/>
-              support@myapp.com
-              </h2>
-            </div>
-        }
-      ></GlobalModalDialog>
-    </Dialog>
-  );
-};
-
-export default GlobalErorrModal;
